+++ src/sagemaker/transformer.py
@@ -52,6 +52,8 @@ class Transformer(object):
                 using the default AWS configuration chain.
             volume_kms_key (str): Optional. KMS key ID for encrypting the volume attached to the ML
                 compute instance (default: None).
+            model (sagemaker.model.Model): A SageMaker Model object, used for SageMaker Model interactions
+                (default: None). If not specified, model object related activities will fail.
         """
         self.model_name = model_name
         self.strategy = strategy
@@ -111,6 +113,12 @@ class Transformer(object):
 
         self.latest_transform_job = _TransformJob.start_new(self, data, data_type, content_type, compression_type,
                                                             split_type)
+
+    def delete_model(self):
+        """Delete a SageMaker Model.
+
+        """
+        self.sagemaker_session.delete_model(self.model_name)
 
     def _retrieve_image_name(self):
         model_desc = self.sagemaker_session.sagemaker_client.describe_model(ModelName=self.model_name)